software/waveshare-eink-pico/requirements.txt
flask
numpy
Pillow
waitress